// src/model/persona.ts
export class Persona {
    constructor() {
      this.id = 0;
      this.nombre = "";
      this.apellido = "";
      this.telefono = "";
      this.direccion = "";
      this.cargo = "";
      this.usuario = "";
      this.contraseña = "";
    }
  
    id: number;
    nombre: string;
    apellido: string;
    telefono: string;
    direccion: string;
    cargo: string;
    usuario: string;
    contraseña: string;
  }
  